pyproject.toml
[tool.poetry]
name = "monnt.py"
version = "0.1.0"
description = ""
authors = ["Thomas Schmitt <t.e.schmitt@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
tortoise-orm = "^0.19.0"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.3.2"
aerich = "^0.6.2"


[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
